chore: move repl_cfg to separate file
